UI and stuff
I did the timer and chits stats on the UI. I made basic functionality to the cage and fish tank. I used the bottles to practice and the code needs to be changed to chits not bottles. Bottles can be yeeted and destroyed. Took off rigidbodies for big furniture. Made a game over screen as a place holder.